fix: MaterialStatus removed as a parameter and added @Service for MachineQueryService.

+++ src/main/java/com/construtech/buildsphere/platform/resourceManagement/interfaces/rest/resources/CreateMaterialResource.java
@@ -7,6 +7,6 @@ public record CreateMaterialResource(
         int amount,
         double totalCost,
         String status,
-        Long projectId
+        int projectId
 ) {
 }
