+++ src/main/java/com/construtech/buildsphere/platform/resourceManagement/interfaces/rest/resources/MachineResource.java
@@ -9,6 +9,6 @@ public record MachineResource(
         LocalDate receptionDate,
         LocalDate endDate,
         double totalCost,
-        Long projectId
+        int projectId
 ) {
 }
